Smoke test

See merge request coccoc/qa-auto!152

+++ testscripts/smoketest/common.py
@@ -171,7 +171,7 @@ def install_coccoc_silentlty_make_coccoc_default_browser(coccoc_installer_name='
     p.communicate()
 
 
-def cleanup(coccoc_update=True):
+def cleanup(coccoc_update=True, firefox=True):
     # Kill all unncessary task
     if coccoc_update:
         WindowsCMD.execute_cmd('taskkill /im CocCocUpdate.exe /f')
@@ -179,7 +179,8 @@ def cleanup(coccoc_update=True):
     WindowsCMD.execute_cmd('taskkill /im MicrosoftEdgeCP.exe /f')
     WindowsCMD.execute_cmd('taskkill /im MicrosoftEdgeCP.exe /f')
     WindowsCMD.execute_cmd('taskkill /im iexplore.exe /f')
-    WindowsCMD.execute_cmd('taskkill /im firefox.exe /f')
+    if firefox:
+        WindowsCMD.execute_cmd('taskkill /im firefox.exe /f')
 
 
 def wait_for_coccoc_install_finish():
@@ -331,12 +332,12 @@ def open_link_from_powershell():
         print("Ignore error code")
 
 
-def get_coccoc_process():
+def get_application_process(application_name='browser'):
     import subprocess
     processes = None
     try:
         processes = subprocess.Popen(["powershell.exe",
-                                      "Get-Process browser"],
+                                      f"Get-Process {application_name}"],
                                      stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
     except:
         print("Ignore error code")
